Support MT9v034 high-resolution image data transmission

+++ include/system/messages/message.hpp
@@ -24,8 +24,7 @@ public:
   void defaultInit();
   void setWholeMessageSizeWithInnerMessage(uint32_t inner_msg_size);
   uint8_t getNumberOfBytesBeforeInnerMessage();
-  uint8_t getInnerMessageSizeBySubtraction();
-
+  uint32_t getInnerMessageSizeBySubtraction();
 };
 
 #endif /* MESSAGE_H */
